Split out components from filters/existing-container
